VGG16/requirements.txt
Flask
tensorflow==2.16.1
numpy
pillow
gunicorn
